x86/boot: Convert early_serial_base to unsigned long
As a preparatory of adding flexible serial I/O accessors, convert
early_serial_base to unsigned long to cover all possible bus addresses
on the system.

No functional change intended.

Link: https://lore.kernel.org/r/20250502123145.4066635-2-andriy.shevchenko@linux.intel.com

+++ arch/x86/boot/boot.h
@@ -274,7 +274,7 @@ int check_knl_erratum(void);
 int validate_cpu(void);
 
 /* early_serial_console.c */
-extern int early_serial_base;
+extern unsigned long early_serial_base;
 void console_init(void);
 
 /* edd.c */
